Set uploads directory permissions in web container too
There is a race condition that occurs depending on the order that the containers
are started. The ownerships on the host of the volume is set to the last
container that comes up. Therefore, before this change, if the web container
came up after the PHP container the uploads directory will be owned by root.

To resolve this, we set the permissions in the web container too. The UID of 82
is standard for Alpine's www-data user[0].

[0] https://github.com/docker-library/php/blob/903540ea7918b5cabed6b32e81f8518f9e6f204f/7.1/fpm/alpine/Dockerfile#L31.